feat: fixed all the norminette errors
All the norminette errors in the project have been
fixed. The ones that remain are from the debug.c
file, the commented if conditions for the exit
builtin and the comment of the PRINT_CMD().
