+++ sources/cleanup/free.c
@@ -6,7 +6,7 @@
 /*   By: mdanish <mdanish@student.42abudhabi.ae>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/04 17:38:52 by maabdull          #+#    #+#             */
-/*   Updated: 2024/08/02 21:33:43 by mdanish          ###   ########.fr       */
+/*   Updated: 2024/08/02 21:56:02 by mdanish          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
