+++ sources/execution/module.mk
@@ -1,2 +1,2 @@
-EXEC_SRCS := $(addprefix $(EXEC_DIR)/, new_exec.c)
+EXEC_SRCS := $(addprefix $(EXEC_DIR)/, new_exec.c execution_utils.c)
 SRCS += $(EXEC_SRCS)
