+++ sources/main.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   main.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: maabdull <maabdull@student.42abudhabi.a    +#+  +:+       +#+        */
+/*   By: mdanish <mdanish@student.42abudhabi.ae>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/31 13:43:49 by maabdull          #+#    #+#             */
-/*   Updated: 2024/08/01 23:34:52 by maabdull         ###   ########.fr       */
+/*   Updated: 2024/08/02 17:46:12 by mdanish          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,9 +38,9 @@ int	main(int argc, char *argv[]__attribute__((unused)), char **env)
 		minishell.output_fd = false;
 		minishell.cmd = NULL;
 		parse(&minishell, line, line);
-		// run_command(&minishell, -1);
+		run_command(&minishell, -1);
 		//! FOR DEBUGGING:
-		PRINT_CMD(minishell.cmd);
+		// PRINT_CMD(minishell.cmd);
 		free_parsing(&minishell);
 	}
 	free_environment(&minishell);
